new Entity is added

+++ HealthCareManagementSystem/src/main/java/com/capgemini/hms/repository/DoctorDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.capgemini.hms.entity.Doctor;
 
-public interface DoctorDao extends JpaRepository<Doctor, Long>{
+public interface DoctorDao extends JpaRepository<Doctor, Integer>{
 
 }
